Get simple color API working
