+++ examples/Pong/Pong.fs
@@ -84,12 +84,12 @@ open Fable.Core.Rust
 let init (_args: array<string>) =
     game
     |> GameBuilder.draw3d (fun model frameTime -> 
-        Graphics.Scene3D.group([|
+        Graphics.Scene3D.material (Graphics.Scene3D.Material.color(0.0f, 1.0f, 0.0f, 1.0f), [|
             Graphics.Scene3D.cylinder() |> Graphics.Scene3D.Transform.translateY -1.0f;
             Graphics.Scene3D.cube()
             |> Graphics.Scene3D.Transform.rotateZ (Math.Angle.degrees (frameTime.tts * 40.0f))
         |])
         |> Graphics.Scene3D.Transform.translateZ ((sin (frameTime.tts * 5.0f)) * 1.0f)
-        )
+    )
     |> GameBuilder.tick tick
     |> Runtime.runGame
