+++ runtime/functor-runtime-common/src/scene3d/mod.rs
@@ -10,6 +10,12 @@ use crate::{
     RenderContext,
 };
 
+mod material_description;
+mod texture_description;
+
+pub use material_description::*;
+pub use texture_description::*;
+
 #[derive(Debug, Clone, Serialize, Deserialize)]
 pub enum Shape {
     Cube,
@@ -18,50 +24,10 @@ pub enum Shape {
 }
 
 #[derive(Debug, Clone, Serialize, Deserialize)]
-pub enum MaterialDescription {
-    #[serde(
-        serialize_with = "serialize_vec4",
-        deserialize_with = "deserialize_vec4"
-    )]
-    Color(Vector4<f32>),
-}
-
-impl MaterialDescription {
-    // TODO: Use color type
-    pub fn color(r: f32, g: f32, b: f32, a: f32) -> MaterialDescription {
-        MaterialDescription::Color(vec4(r, g, b, a))
-    }
-}
-
-impl MaterialDescription {
-    pub fn get(&self, context: &RenderContext) -> Box<dyn Material> {
-        match self {
-            MaterialDescription::Color(c) => {
-                // TODO: Load from cache of assets
-                let mut color_material = ColorMaterial::create(*c);
-                color_material.initialize(&context);
-                color_material
-            }
-        }
-    }
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
 pub enum SceneObject {
     Geometry(Shape),
     Material(MaterialDescription, Vec<Scene3D>),
     Group(Vec<Scene3D>),
-}
-
-#[derive(Debug)]
-pub enum Texture2DHandle {
-    File(String),
-}
-
-impl Texture2DHandle {
-    pub fn file(s: LrcStr) -> Texture2DHandle {
-        Texture2DHandle::File(s.to_string())
-    }
 }
 
 #[derive(Debug, Clone, Serialize, Deserialize)]
@@ -224,22 +190,6 @@ impl Scene3D {
             }
         }
     }
-}
-
-fn serialize_vec4<S>(v: &Vector4<f32>, serializer: S) -> Result<S::Ok, S::Error>
-where
-    S: serde::Serializer,
-{
-    let array: [f32; 4] = [v.x, v.y, v.z, v.w];
-    array.serialize(serializer)
-}
-
-fn deserialize_vec4<'de, D>(deserializer: D) -> Result<Vector4<f32>, D::Error>
-where
-    D: serde::Deserializer<'de>,
-{
-    let array = <[f32; 4]>::deserialize(deserializer)?;
-    Ok(Vector4::new(array[0], array[1], array[2], array[3]))
 }
 
 fn serialize_matrix<S>(matrix: &Matrix4<f32>, serializer: S) -> Result<S::Ok, S::Error>
