+++ runtime/functor-runtime-common/src/scene3d/texture_description.rs
@@ -1,0 +1,12 @@
+use fable_library_rust::String_::LrcStr;
+
+#[derive(Debug)]
+pub enum TextureDescription {
+    File(String),
+}
+
+impl TextureDescription {
+    pub fn file(s: LrcStr) -> TextureDescription {
+        TextureDescription::File(s.to_string())
+    }
+}
