+++ src/Functor.Game/Scene3D.fs
@@ -3,7 +3,8 @@ namespace Graphics
 open Fable.Core
 [<Erase; Emit("functor_runtime_common::Scene3D")>] type Scene3D = | Noop
 
-[<Erase; Emit("functor_runtime_common::Texture2DHandle")>] type Texture2D = | Noop
+[<Erase; Emit("functor_runtime_common::TextureDescription")>] type Texture = | Noop
+[<Erase; Emit("functor_runtime_common::MaterialDescription")>] type Material = | Noop
 
 module Scene3D =
 
@@ -19,9 +20,17 @@ module Scene3D =
     [<Emit("functor_runtime_common::Scene3D::group($0)")>]
     let group (items: Scene3D[]): Scene3D = nativeOnly
 
+    [<Emit("functor_runtime_common::Scene3D::material($0, $1)")>]
+    let material (material: Material, items: Scene3D[]): Scene3D = nativeOnly
+    
+    module Material =
+        [<Emit("functor_runtime_common::MaterialDescription::color($0, $1, $2, $3)")>]
+        let color (r: float32, b: float32, g: float32, a: float32): Material = nativeOnly
+
+
     module Texture2D =
-        [<Emit("functor_runtime_common::Texture2DHandle::file($0)")>]
-        let file (str: string): Texture2D = nativeOnly
+        [<Emit("functor_runtime_common::TextureDescription::file($0)")>]
+        let file (str: string): Texture = nativeOnly
 
     module Transform = 
         [<Emit("functor_runtime_common::Scene3D::translate_x($1, $0)")>]
